requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
black==24.4.2
pytest
pytest-cov
pytest-django
drf-spectacular
drf-spectacular-sidecar
